+++ src/core/tools/utils.ml
@@ -93,6 +93,28 @@ let read_all filename =
   close_in channel;
   Strings.Mutable.to_string contents
 
+let copy ?(mode = [Open_wronly; Open_creat; Open_trunc]) ?(perms = 0o660) src
+    dst =
+  let oc = open_out_gen mode perms dst in
+  Fun.protect
+    ~finally:(fun () -> close_out_noerr oc)
+    (fun () ->
+      set_binary_mode_out oc true;
+      let ic = open_in_bin src in
+      Fun.protect
+        ~finally:(fun () -> close_in_noerr ic)
+        (fun () ->
+          let len = 4096 in
+          let buf = Bytes.create len in
+          let rec f () =
+            match input ic buf 0 len with
+              | 0 -> ()
+              | n ->
+                  output_substring oc (Bytes.unsafe_to_string buf) 0 n;
+                  f ()
+          in
+          f ()))
+
 (* Drop the first [len] bytes. *)
 let buffer_drop buffer len =
   let size = Buffer.length buffer in